+++ CHANGELOG.md
@@ -7,6 +7,7 @@ ChangeLog
 
 - UPGRADE: upgrade NPM dependencies
 - IMPROVEMENT: allow nodes to publish dashboard info to other nodes, too
+- IMPROVEMENT: add "speex" noise suppression node
 
 1.4.5 (2025-08-07)
 ------------------